+++ src/index.ts
@@ -1,10 +1,10 @@
-import { Post } from "./entities/Post";
+import { Post as PostEntity } from "./entities/post.entity";
 import { MikroORM } from "@mikro-orm/core";
 import microConfig from "./mikro-orm.config";
 import express from "express";
 import { ApolloServer } from "apollo-server-express";
 import { buildSchema } from "type-graphql";
-import { NumberResolver } from "./resolvers/number";
+import { NumberResolver } from "./resolvers/number.resolver";
 import { PostResolver } from "./resolvers/post.resolver";
 
 const main = async () => {
@@ -23,7 +23,7 @@ const main = async () => {
 
     apolloServer.applyMiddleware({ app });
 
-    const posts = await orm.em.find(Post, {});
+    const posts = await orm.em.find(PostEntity, {});
     app.get("/", (_, res) => res.send(posts));
     app.listen(4000, () =>
         console.log("Server started at http://localhost:4000")
