+++ src/mikro-orm.config.ts
@@ -1,4 +1,4 @@
-import { Post } from "./entities/Post";
+import { Post } from "./entities/post.entity";
 import { __prod__ } from "./constants";
 import { MikroORM } from "@mikro-orm/core";
 import path from "path";
